+++ backend/src/api/controllers/result.ts
@@ -1,11 +1,6 @@
 import * as ResultDAL from "../../dal/result";
 import * as PublicDAL from "../../dal/public";
-import {
-  getCurrentDayTimestamp,
-  getStartOfDayTimestamp,
-  isDevEnvironment,
-  replaceObjectId,
-} from "../../utils/misc";
+import { isDevEnvironment, replaceObjectId } from "../../utils/misc";
 import objectHash from "object-hash";
 import Logger from "../../utils/logger";
 import "dotenv/config";
@@ -53,6 +48,10 @@ import {
 } from "@monkeytype/contracts/schemas/results";
 import { Mode } from "@monkeytype/contracts/schemas/shared";
 import { mapRange, roundTo2, stdDev } from "@monkeytype/util/numbers";
+import {
+  getCurrentDayTimestamp,
+  getStartOfDayTimestamp,
+} from "@monkeytype/util/date-and-time";
 
 try {
   if (!anticheatImplemented()) throw new Error("undefined");
