+++ backend/src/dal/user.ts
@@ -9,7 +9,7 @@ import {
   type UpdateFilter,
   type Filter,
 } from "mongodb";
-import { flattenObjectDeep, isToday, isYesterday } from "../utils/misc";
+import { flattenObjectDeep } from "../utils/misc";
 import { getCachedConfiguration } from "../init/configuration";
 import { getDayOfYear } from "date-fns";
 import { UTCDate } from "@date-fns/utc";
@@ -32,6 +32,7 @@ import {
 import { addImportantLog } from "./logs";
 import { Result as ResultType } from "@monkeytype/contracts/schemas/results";
 import { Configuration } from "@monkeytype/contracts/schemas/configuration";
+import { isToday, isYesterday } from "@monkeytype/util/date-and-time";
 
 const SECONDS_PER_HOUR = 3600;
