+++ backend/src/queues/later-queue.ts
@@ -1,8 +1,11 @@
 import LRUCache from "lru-cache";
 import Logger from "../utils/logger";
 import { MonkeyQueue } from "./monkey-queue";
-import { getCurrentDayTimestamp, getCurrentWeekTimestamp } from "../utils/misc";
 import { ValidModeRule } from "@monkeytype/contracts/schemas/configuration";
+import {
+  getCurrentDayTimestamp,
+  getCurrentWeekTimestamp,
+} from "@monkeytype/util/date-and-time";
 
 const QUEUE_NAME = "later";
