+++ frontend/src/ts/elements/profile.ts
@@ -4,7 +4,7 @@ import { differenceInDays } from "date-fns/differenceInDays";
 import * as Misc from "../utils/misc";
 import * as Numbers from "@monkeytype/util/numbers";
 import * as Levels from "../utils/levels";
-import * as DateTime from "../utils/date-and-time";
+import * as DateTime from "@monkeytype/util/date-and-time";
 import { getHTMLById } from "../controllers/badge-controller";
 import { throttle } from "throttle-debounce";
 import * as ActivePage from "../states/active-page";
@@ -13,6 +13,7 @@ import { getHtmlByUserFlags } from "../controllers/user-flag-controller";
 import Format from "../utils/format";
 import { UserProfile, RankAndCount } from "@monkeytype/contracts/schemas/users";
 import { abbreviateNumber, convertRemToPixels } from "../utils/numbers";
+import { secondsToString } from "../utils/date-and-time";
 
 type ProfileViewPaths = "profile" | "account";
 type UserProfileOrSnapshot = UserProfile | MonkeyTypes.Snapshot;
@@ -228,7 +229,7 @@ export async function update(
   typingStatsEl
     .find(".timeTyping .value")
     .text(
-      DateTime.secondsToString(
+      secondsToString(
         Math.round(profile.typingStats?.timeTyping ?? 0),
         true,
         true
