+++ frontend/src/ts/utils/date-and-time.ts
@@ -1,64 +1,6 @@
 import { roundTo2 } from "@monkeytype/util/numbers";
 
 /**
- * Returns the current day's timestamp adjusted by the hour offset.
- * @param hourOffset The offset in hours. Default is 0.
- * @returns The timestamp of the start of the current day adjusted by the hour offset.
- */
-export function getCurrentDayTimestamp(hourOffset = 0): number {
-  const offsetMilis = hourOffset * MILISECONDS_IN_HOUR;
-  const currentTime = Date.now();
-  return getStartOfDayTimestamp(currentTime, offsetMilis);
-}
-
-const MILISECONDS_IN_HOUR = 3600000;
-const MILLISECONDS_IN_DAY = 86400000;
-
-/**
- * Returns the timestamp of the start of the day for the given timestamp adjusted by the offset.
- * @param timestamp The timestamp for which to get the start of the day.
- * @param offsetMilis The offset in milliseconds. Default is 0.
- * @returns The timestamp of the start of the day for the given timestamp adjusted by the offset.
- */
-export function getStartOfDayTimestamp(
-  timestamp: number,
-  offsetMilis = 0
-): number {
-  return timestamp - ((timestamp - offsetMilis) % MILLISECONDS_IN_DAY);
-}
-
-/**
- * Checks if the given timestamp is from yesterday, adjusted by the hour offset.
- * @param timestamp The timestamp to check.
- * @param hourOffset The offset in hours. Default is 0.
- * @returns True if the timestamp is from yesterday, false otherwise.
- */
-export function isYesterday(timestamp: number, hourOffset = 0): boolean {
-  const offsetMilis = hourOffset * MILISECONDS_IN_HOUR;
-  const yesterday = getStartOfDayTimestamp(
-    Date.now() - MILLISECONDS_IN_DAY,
-    offsetMilis
-  );
-  const date = getStartOfDayTimestamp(timestamp, offsetMilis);
-
-  return yesterday === date;
-}
-
-/**
- * Checks if the given timestamp is from today, adjusted by the hour offset.
- * @param timestamp The timestamp to check.
- * @param hourOffset The offset in hours. Default is 0.
- * @returns True if the timestamp is from today, false otherwise.
- */
-export function isToday(timestamp: number, hourOffset = 0): boolean {
-  const offsetMilis = hourOffset * MILISECONDS_IN_HOUR;
-  const today = getStartOfDayTimestamp(Date.now(), offsetMilis);
-  const date = getStartOfDayTimestamp(timestamp, offsetMilis);
-
-  return today === date;
-}
-
-/**
  * Converts seconds to a human-readable string representation of time.
  * @param sec The number of seconds to convert.
  * @param alwaysShowMinutes Whether to always show minutes, even if the value is 0. Default is false.
